Remove bvt tag (#5967)
Remove bvt tag

Approved by: @fengttt, @daviszhen, @aressu1985, @m-schen, @nnsgmsone

+++ pkg/sql/colexec/auto_incr.go
@@ -303,6 +303,9 @@ func getCurrentIndex(param *AutoIncrParam, colName string, mp *mpool.MPool) (uin
 
 func updateAutoIncrTable(param *AutoIncrParam, curNum uint64, name string, mp *mpool.MPool) error {
 	bat, _ := GetDeleteBatch(param.rel, param.ctx, name, mp)
+	if bat == nil {
+		return moerr.NewInternalError("the deleted batch is nil")
+	}
 	bat.SetZs(bat.GetVector(0).Length(), mp)
 	err := param.rel.Delete(param.ctx, bat, AUTO_INCR_TABLE_COLNAME[0])
 	if err != nil {
@@ -454,6 +457,9 @@ func DeleteAutoIncrCol(eg engine.Engine, ctx context.Context, rel engine.Relatio
 				continue
 			}
 			bat, _ := GetDeleteBatch(rel2, ctx, tableID+"_"+d.Attr.Name, proc.Mp())
+			if bat == nil {
+				return moerr.NewInternalError("the deleted batch is nil")
+			}
 			if err = rel2.Delete(ctx, bat, AUTO_INCR_TABLE_COLNAME[0]); err != nil {
 				bat.Clean(proc.Mp())
 				if err2 := RolllbackTxn(eg, txn, ctx); err2 != nil {
@@ -500,6 +506,9 @@ func MoveAutoIncrCol(eg engine.Engine, ctx context.Context, tblName string, db e
 			}
 
 			bat, currentNum := GetDeleteBatch(autoRel, ctx, oldTableID+"_"+d.Attr.Name, proc.Mp())
+			if bat == nil {
+				return moerr.NewInternalError("the deleted batch is nil")
+			}
 			if err = autoRel.Delete(ctx, bat, AUTO_INCR_TABLE_COLNAME[0]); err != nil {
 				if err2 := RolllbackTxn(eg, txn, ctx); err2 != nil {
 					return err2
@@ -550,6 +559,9 @@ func ResetAutoInsrCol(eg engine.Engine, ctx context.Context, tblName string, db 
 				continue
 			}
 			bat, _ := GetDeleteBatch(autoRel, ctx, tableID+"_"+d.Attr.Name, proc.Mp())
+			if bat == nil {
+				return moerr.NewInternalError("the deleted batch is nil")
+			}
 			if err = autoRel.Delete(ctx, bat, AUTO_INCR_TABLE_COLNAME[0]); err != nil {
 				if err2 := RolllbackTxn(eg, txn, ctx); err2 != nil {
 					return err2
